kasih mata login fail

+++ database/migrations/2024_08_07_061105_add_remember_token_to_users_table.php
@@ -6,17 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    public function up(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
+    public function up()
+{
+    Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'remember_token')) {
             $table->string('remember_token', 100)->nullable()->after('password');
-        });
-    }
+        }
+    });
+}
 
-    public function down(): void
-    {
-        Schema::table('users', function (Blueprint $table) {
+public function down()
+{
+    Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'remember_token')) {
             $table->dropColumn('remember_token');
-        });
-    }
+        }
+    });
+}
 };
